requirements.txt
streamlit
spacy
pandas
scikit-learn
PyPDF2
python-docx